Reimport tests from WC2-132 branch
