+++ iaso/tests/api/workflows/test_workflows_followups.py
@@ -1,0 +1,5 @@
+from iaso.tests.api.workflows.base import BaseWorkflowsAPITestCase
+
+
+class WorkflowsFollowupsAPITestCase(BaseWorkflowsAPITestCase):
+    pass
