+++ iaso/tests/api/workflows/test_workflows_mobile.py
@@ -1,0 +1,74 @@
+import jsonschema
+from iaso.tests.api.workflows.base import BaseWorkflowsAPITestCase
+
+
+class WorkflowsMobileAPITestCase(BaseWorkflowsAPITestCase):
+    def test_mobile_api_without_app_id(self):
+        self.client.force_authenticate(self.blue_adult_1)
+
+        response = self.client.get("/api/mobile/workflows/")
+
+        self.assertJSONResponse(response, 404)
+
+        assert response.data == "No app_id provided"
+
+    def test_mobile_api_with_nonexisting_app_id(self):
+        self.client.force_authenticate(self.blue_adult_1)
+
+        response = self.client.get("/api/mobile/workflows/?app_id=wrong_app_id")
+
+        self.assertJSONResponse(response, 404)
+
+        assert (
+            response.data
+            == "User not found in Projects for this app id or project not found"
+        )
+
+    def test_mobile_api_with_nonaccessible_app_id(self):
+        self.client.force_authenticate(self.blue_adult_1)
+
+        response = self.client.get("/api/mobile/workflows/?app_id=red.adults.project")
+
+        self.assertJSONResponse(response, 404)
+
+        assert (
+            response.data
+            == "User not found in Projects for this app id or project not found"
+        )
+
+    def test_mobile_api_ok(self):
+        self.client.force_authenticate(self.blue_adult_1)
+
+        set_tl_schema = {
+            "type": "object",
+            "properties": {
+                "workflows": {
+                    "type": "array",
+                    "items": {
+                        "type": "object",
+                        "properties": {
+                            "status": {"type": "string"},
+                            "created_at": {"type": "number"},
+                            "updated_at": {"type": "number"},
+                            "version_id": {"type": "number"},
+                            "entity_type_id": {"type": "number"},
+                            "name": {"type": "string"},
+                            "changes": {"type": "array"},
+                            "follow_ups": {"type": "array"},
+                        },
+                    },
+                }
+            },
+            "required": ["workflows"],
+        }
+
+        response = self.client.get("/api/mobile/workflows/?app_id=blue.adults.project")
+
+        self.assertJSONResponse(response, 200)
+
+        try:
+            jsonschema.validate(instance=response.data, schema=set_tl_schema)
+        except jsonschema.exceptions.ValidationError as ex:
+            self.fail(msg=str(ex))
+
+        assert len(response.data["workflows"]) == 1
